refactor home screen viewmodel and move design inside screen content
