+++ feature/home/presentation/src/main/java/com/sanaa/presentation/components/MediaListSectionContent.kt
@@ -12,17 +12,17 @@ import androidx.compose.ui.unit.dp
 import com.sanaa.designsystem.design_system.component.chips.ToggleableChip
 import com.sanaa.feature.home.presentation.R
 import com.sanaa.presentation.state.GenreUiState
-import com.sanaa.presentation.state.MediaItem
+import com.sanaa.presentation.state.MediaItemUiState
 
 @Composable
 fun MediaListSectionContent(
     genres: List<GenreUiState>,
-    mediaList: List<MediaItem>,
+    mediaList: List<MediaItemUiState>,
     selectedGenreId: Int?,
     onGenreClick: (Int?) -> Unit,
-    onMediaClick: (MediaItem) -> Unit,
+    onMediaClick: (MediaItemUiState) -> Unit,
     modifier: Modifier = Modifier,
-    onSaveIconClick: (MediaItem) -> Unit = {},
+    onSaveIconClick: (MediaItemUiState) -> Unit = {},
     isScrollEnabled: Boolean = true,
 ) {
     Column(
