+++ feature/home/presentation/src/main/java/com/sanaa/presentation/screen/homeScreen/HomeScreenUiState.kt
@@ -1,20 +1,21 @@
 package com.sanaa.presentation.screen.homeScreen
 
 import androidx.paging.PagingData
+import com.sanaa.presentation.components.SnackData
 import com.sanaa.presentation.state.GenreUiState
-import com.sanaa.presentation.state.MediaItem
+import com.sanaa.presentation.state.MediaItemUiState
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
 
 
 data class HomeScreenUiState(
-    val popularMedia: List<MediaItem> = emptyList(),
-    val topRatingMedia: List<MediaItem> = emptyList(),
-    val continueWatchingMedia: List<MediaItem> = emptyList(),
-    val upcomingMovies: Flow<PagingData<MediaItem>> = flowOf(PagingData.empty()),
+    val popularMedia: List<MediaItemUiState> = emptyList(),
+    val topRatingMedia: List<MediaItemUiState> = emptyList(),
+    val continueWatchingMedia: List<MediaItemUiState> = emptyList(),
+    val upcomingMovies: Flow<PagingData<MediaItemUiState>> = flowOf(PagingData.empty()),
     val movieGenres: List<GenreUiState> = emptyList(),
     val movieSelectedGenreId: Int? = null,
-    val showBottomSheet: Boolean = false,
+    val showLoginBottomSheet: Boolean = false,
     val isLoadingGenre: Boolean = false,
     val isLoadingPopular: Boolean = false,
     val isLoadingTopRated: Boolean = false,
@@ -23,7 +24,8 @@ data class HomeScreenUiState(
     val isNoInternetConnection: Boolean = false,
     val userIsLoggedIn: Boolean = false,
     val showSaveToListBottomSheet: Boolean = false,
-    val selectedMediaToSave: MediaItem? = null,
+    val selectedMediaToSave: MediaItemUiState? = null,
     val showAddListBottomSheet: Boolean = false,
-    val selectedMediaId: Long = 0L
+    val selectedMediaId: Long = 0L,
+    val snackBarData: SnackData? = null
 )
