+++ feature/home/presentation/src/main/java/com/sanaa/presentation/screen/mediaTabScreen/topRatingScreen/TopRatedMediaScreenUiState.kt
@@ -2,15 +2,15 @@ package com.sanaa.presentation.screen.mediaTabScreen.topRatingScreen
 
 import androidx.paging.PagingData
 import com.sanaa.presentation.state.GenreUiState
-import com.sanaa.presentation.state.MediaItem
+import com.sanaa.presentation.state.MediaItemUiState
 import com.sanaa.presentation.state.MediaTypeUi
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
 
 data class TopRatedMediaScreenUiState(
     val selectedMediaTypeUi: MediaTypeUi = MediaTypeUi.MOVIE,
-    val movieList: Flow<PagingData<MediaItem>> = flowOf(PagingData.empty()),
-    val tvShowList: Flow<PagingData<MediaItem>> = flowOf(PagingData.empty()),
+    val movieList: Flow<PagingData<MediaItemUiState>> = flowOf(PagingData.empty()),
+    val tvShowList: Flow<PagingData<MediaItemUiState>> = flowOf(PagingData.empty()),
     val movieGenres: List<GenreUiState> = emptyList(),
     val tvShowGenres: List<GenreUiState> = emptyList(),
     val movieSelectedGenreId: Int? = null,
@@ -21,5 +21,5 @@ data class TopRatedMediaScreenUiState(
     val userIsLoggedIn: Boolean = false,
     val showSaveToListBottomSheet: Boolean = false,
     val showAddListBottomSheet: Boolean = false,
-    val selectedMediaToSave: MediaItem? = null,
+    val selectedMediaToSave: MediaItemUiState? = null,
 )
