+++ feature/home/presentation/src/main/java/com/sanaa/presentation/screen/mediaTabScreen/watchingHistoryScreen/WatchingMediaHistoryScreenInteractionListener.kt
@@ -1,6 +1,6 @@
 package com.sanaa.presentation.screen.mediaTabScreen.watchingHistoryScreen
 
-import com.sanaa.presentation.state.MediaItem
+import com.sanaa.presentation.state.MediaItemUiState
 import com.sanaa.presentation.state.MediaTypeUi
 
 interface WatchingMediaHistoryScreenInteractionListener {
@@ -8,7 +8,7 @@ interface WatchingMediaHistoryScreenInteractionListener {
     fun onMovieGenreClick(id: Int?)
     fun onTvShowGenreClick(id: Int?)
     fun onMediaClick(id: Int, mediaTypeUi: MediaTypeUi)
-    fun onSaveIconClick(media: MediaItem)
+    fun onSaveIconClick(media: MediaItemUiState)
     fun onBackClick()
     fun onRetryClick()
 }
