+++ feature/home/presentation/src/main/java/com/sanaa/presentation/screen/trendingMediaScreen/TrendingMediaScreenInteractionListener.kt
@@ -1,11 +1,11 @@
 package com.sanaa.presentation.screen.trendingMediaScreen
 
-import com.sanaa.presentation.state.MediaItem
+import com.sanaa.presentation.state.MediaItemUiState
 
 interface MediaListScreenInteractionListener {
     fun onGenreClick(id: Int?)
     fun onMediaClick(id: Int)
-    fun onSaveIconClick(media: MediaItem)
+    fun onSaveIconClick(media: MediaItemUiState)
     fun onSaveToListSuccess()
     fun onSaveToListFailure()
     fun onBackClick()
