+++ feature/home/presentation/src/main/java/com/sanaa/presentation/screen/trendingMediaScreen/TrendingMediaScreenUiState.kt
@@ -2,12 +2,12 @@ package com.sanaa.presentation.screen.trendingMediaScreen
 
 import androidx.paging.PagingData
 import com.sanaa.presentation.state.GenreUiState
-import com.sanaa.presentation.state.MediaItem
+import com.sanaa.presentation.state.MediaItemUiState
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
 
 data class TrendingMediaScreenUiState (
-    val mediaList: Flow<PagingData<MediaItem>> = flowOf(PagingData.empty()),
+    val mediaList: Flow<PagingData<MediaItemUiState>> = flowOf(PagingData.empty()),
     val genreList: List<GenreUiState> = emptyList(),
     val selectedGenreId: Int? = null,
     val isLoading: Boolean = false,
