+++ feature/home/presentation/src/main/java/com/sanaa/presentation/screen/trendingMediaScreen/trendingMoviesScreen/TrendingMoviesScreenViewModel.kt
@@ -11,7 +11,7 @@ import com.sanaa.presentation.screen.trendingMediaScreen.MediaListScreenInteract
 import com.sanaa.presentation.screen.trendingMediaScreen.TrendingMediaScreenEffect
 import com.sanaa.presentation.screen.trendingMediaScreen.TrendingMediaScreenUiState
 import com.sanaa.presentation.state.GenreUiState
-import com.sanaa.presentation.state.MediaItem
+import com.sanaa.presentation.state.MediaItemUiState
 import com.sanaa.presentation.state.mapper.toState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import entity.Movie
@@ -77,7 +77,7 @@ class TrendingMoviesScreenViewModel @Inject constructor(
         )
     }
 
-    private fun loadMoviesOperation(): Flow<PagingData<MediaItem>> {
+    private fun loadMoviesOperation(): Flow<PagingData<MediaItemUiState>> {
         return createPagingFlow(
             pagingSourceFactory = { createMoviesPagingSource() },
             mapper = Movie::toState
@@ -88,7 +88,7 @@ class TrendingMoviesScreenViewModel @Inject constructor(
         }.cachedIn(viewModelScope)
     }
 
-    private fun onLoadMoviesSuccess(pagingData: PagingData<MediaItem>) {
+    private fun onLoadMoviesSuccess(pagingData: PagingData<MediaItemUiState>) {
         updateState {
             copy(
                 mediaList = flowOf(pagingData),
@@ -108,7 +108,7 @@ class TrendingMoviesScreenViewModel @Inject constructor(
         emitEffect(TrendingMediaScreenEffect.NavigateToMediaDetails(id))
     }
 
-    override fun onSaveIconClick(media: MediaItem) {
+    override fun onSaveIconClick(media: MediaItemUiState) {
         if (!state.value.userIsLoggedIn) {
             updateState { copy(showLoginBottomSheet = true) }
             return
