+++ feature/home/presentation/src/main/java/com/sanaa/presentation/screen/trendingMediaScreen/trendingTvShowScreen/TrendingTvShowsScreenViewModel.kt
@@ -8,7 +8,7 @@ import com.sanaa.presentation.screen.trendingMediaScreen.MediaListScreenInteract
 import com.sanaa.presentation.screen.trendingMediaScreen.TrendingMediaScreenEffect
 import com.sanaa.presentation.screen.trendingMediaScreen.TrendingMediaScreenUiState
 import com.sanaa.presentation.state.GenreUiState
-import com.sanaa.presentation.state.MediaItem
+import com.sanaa.presentation.state.MediaItemUiState
 import com.sanaa.presentation.state.mapper.toState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import entity.TvSeries
@@ -70,14 +70,14 @@ class TrendingTvShowsScreenViewModel @Inject constructor(
         )
     }
 
-    private fun loadTvShowsOperation(): Flow<PagingData<MediaItem>> {
+    private fun loadTvShowsOperation(): Flow<PagingData<MediaItemUiState>> {
         return createPagingFlow(
             pagingSourceFactory = { createTvShowsPagingSource() },
             mapper = TvSeries::toState
         )
     }
 
-    private fun onLoadTvShowsSuccess(pagingData: PagingData<MediaItem>) {
+    private fun onLoadTvShowsSuccess(pagingData: PagingData<MediaItemUiState>) {
         updateState {
             copy(
                 mediaList = flowOf(pagingData),
@@ -99,7 +99,7 @@ class TrendingTvShowsScreenViewModel @Inject constructor(
         emitEffect(TrendingMediaScreenEffect.NavigateToMediaDetails(id))
     }
 
-    override fun onSaveIconClick(media: MediaItem) {
+    override fun onSaveIconClick(media: MediaItemUiState) {
         if (!state.value.userIsLoggedIn) {
             updateState {
                 copy(
