+++ feature/home/presentation/src/main/java/com/sanaa/presentation/state/mapper/MediaMapper.kt
@@ -2,7 +2,7 @@ package com.sanaa.presentation.state.mapper
 
 import android.annotation.SuppressLint
 import com.sanaa.presentation.bottomsheet.saveToListBottomsheet.PlaylistUiItem
-import com.sanaa.presentation.state.MediaItem
+import com.sanaa.presentation.state.MediaItemUiState
 import com.sanaa.presentation.state.MediaTypeUi
 import entity.MediaHistoryItem
 import entity.Movie
@@ -11,7 +11,7 @@ import usecase.custom_list.custom_list_param.SavedList
 import usecase.search.search_param.MediaType
 
 @SuppressLint("DefaultLocale")
-fun Movie.toState(): MediaItem = MediaItem(
+fun Movie.toState(): MediaItemUiState = MediaItemUiState(
     id = id,
     title = title,
     imageUrl = posterImageUrl,
@@ -21,7 +21,7 @@ fun Movie.toState(): MediaItem = MediaItem(
 )
 
 @SuppressLint("DefaultLocale")
-fun TvSeries.toState(): MediaItem = MediaItem(
+fun TvSeries.toState(): MediaItemUiState = MediaItemUiState(
     id = id,
     title = title,
     imageUrl = posterImageUrl,
@@ -29,7 +29,7 @@ fun TvSeries.toState(): MediaItem = MediaItem(
     mediaTypeUi = MediaTypeUi.TV_SHOW
 )
 
-fun MediaHistoryItem.toState(): MediaItem = MediaItem(
+fun MediaHistoryItem.toState(): MediaItemUiState = MediaItemUiState(
     id = id,
     title = "",
     imageUrl = posterImageUrl,
