+++ feature/home/presentation/src/test/java/com/sanaa/presentation/screen/trendingMediaScreen/trendingMoviesScreen/TrendingMoviesScreenViewModelTest.kt
@@ -3,7 +3,7 @@ package com.sanaa.presentation.screen.trendingMediaScreen.trendingMoviesScreen
 import app.cash.turbine.test
 import com.google.common.truth.Truth.assertThat
 import com.sanaa.presentation.screen.trendingMediaScreen.TrendingMediaScreenEffect
-import com.sanaa.presentation.state.MediaItem
+import com.sanaa.presentation.state.MediaItemUiState
 import com.sanaa.presentation.state.MediaTypeUi
 import com.sanaa.presentation.state.mapper.toState
 import entity.Genre
@@ -225,7 +225,7 @@ class TrendingMoviesScreenViewModelTest {
                 rating = 0
             )
         )
-        val media = MediaItem(
+        val media = MediaItemUiState(
             id = 1,
             title = "Test Movie",
             imageUrl = "https://example.com/image.jpg",
